+++ lilith/src/main/resources/logbackCLI.groovy
@@ -1,0 +1,13 @@
+import ch.qos.logback.classic.encoder.PatternLayoutEncoder
+import ch.qos.logback.core.ConsoleAppender
+
+import static ch.qos.logback.classic.Level.WARN
+
+context.name = 'Lilith'
+appender('CONSOLE', ConsoleAppender) {
+  encoder(PatternLayoutEncoder) {
+    pattern = '%-5level - %msg%n'
+  }
+}
+root(WARN, ['CONSOLE'])
+